another fix
changed stful to stless widgets, created provider file
