+++ lib/main.dart
@@ -2,12 +2,11 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
 import 'package:easy_localization/easy_localization.dart';
-import 'package:useful_useless_app/global/auth/login_page.dart';
 import 'package:useful_useless_app/ui/home_screen.dart';
 
 import 'package:provider/provider.dart';
 export 'package:easy_localization_loader/src/json_asset_loader.dart';
-
+import 'package:useful_useless_app/src/core/provider/provider.dart';
 
 Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
