+++ lib/ui/global/loader_widget.dart
@@ -1,15 +1,9 @@
 import 'package:flutter/material.dart';
 
 
-class Loader extends StatefulWidget {
-  @override
-  _LoaderState createState() => _LoaderState();
-}
-
-class _LoaderState extends State<Loader> {
-
+class Loader extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return CircularProgressIndicator(); //example
+    return CircularProgressIndicator();
   }
 }
