+++ pubspec.yaml
@@ -34,7 +34,7 @@ dependencies:
   provider: ^4.3.2+2
 
 
-  firebase_auth: ^0.18.0+1
+  firebase_auth: ^0.18.1+1
   google_sign_in: ^4.5.4
